+++ CHANGES.md
@@ -1,4 +1,5 @@
 **Added:**
+* ability to restart application
 * more control icons
 
 **Changed:**